поделил эндпойнты на 2 контроллера
